Move the xcomponent bindings to the top level
