+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+# v0.2.0 (unreleased)
+
+## Breaking
+
+- Renamed and moved the xcomponent module to the top-level  (from ace/xcomponent/native_interface_xcomponent)
