+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "ohos-sys"
-version = "0.1.0"
+version = "0.2.0-beta1"
 edition = "2021"
 description = "Bindings to the native API of OpenHarmony OS"
 license = "Apache-2.0"
