--- src/ace/mod.rs
@@ -1,2 +1,0 @@
-//! Bindings for `libace_ndk.z.so`
-pub mod xcomponent;
