--- src/ace/xcomponent/mod.rs
@@ -1,1 +1,0 @@
-pub mod native_interface_xcomponent;
