+++ src/lib.rs
@@ -1,5 +1,5 @@
-pub mod ace;
 pub mod hilog;
 pub mod napi;
 pub mod native_buffer;
 pub mod native_window;
+pub mod xcomponent;
